+++ src/main/java/uz/tsx/controller/UserController.java
@@ -43,7 +43,7 @@ public class UserController {
     @PreAuthorize("hasAnyRole('ADMIN')")
     @Operation(summary = "This method for get", description = "This method is used to get how many points the admin user has scored")
     @GetMapping(value = "/info/{id}")
-    public HttpResponse<Object> getUserInformation(@PathVariable Integer id) {
+    public HttpResponse<Object> getUserInformation(@PathVariable Long id) {
 
         UserInterface userInformation = service.getById(id);
         UserDto userDto = UserConvert.from(userInformation);
@@ -111,7 +111,7 @@ public class UserController {
     @PreAuthorize("hasRole('ADMIN')")
     @Operation(summary = "This method for update by id", description = "This method updates the user's data")
     @PutMapping("/update/{id}")
-    public HttpResponse<Object> userUpdateId(@PathVariable Integer id, @RequestBody UserUpdateRequestDto userUpdate) {
+    public HttpResponse<Object> userUpdateId(@PathVariable Long id, @RequestBody UserUpdateRequestDto userUpdate) {
 
         UserEntity updateUser = userUpdate.toEntity();
         Boolean isUpdateUser = service.updateById(updateUser, id);
@@ -128,7 +128,7 @@ public class UserController {
     @Operation(summary = "This user for delete", description = "This method is designed to delete a user by ID")
     @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/delete/{id}")
-    public HttpResponse<Object> userDelete(@PathVariable Integer id) {
+    public HttpResponse<Object> userDelete(@PathVariable Long id) {
 
         service.delete(id);
         return HttpResponse.build()
