+++ src/main/java/uz/tsx/service/UserService.java
@@ -7,5 +7,5 @@ public interface UserService extends BaseInterface<UserInterface>{
 
     Boolean updateMe(UserEntity userUpdate, String attachId);
 
-    Boolean updateById(UserEntity userUpdate, Integer id);
+    Boolean updateById(UserEntity userUpdate, Long id);
 }
